+++ react/ui-components/src/atoms/SearchComponent.js
@@ -32,7 +32,6 @@ const SearchComponent = ({ uiConfig, header = "", screenType = "search", fullCon
   const [session,setSession,clearSession] = browserSession || []
   const buttonWrapperRef = useRef(null);
   const [addMargin, setAddMargin] = useState(false);
-  const [sortOrder, setSortOrder] = useState(uiConfig?.sortConfig?.initialSortOrder || 'asc');
   
   if (fullConfig?.postProcessResult){
     //conditions can be added while calling postprocess function to pass different params
@@ -114,7 +113,7 @@ const SearchComponent = ({ uiConfig, header = "", screenType = "search", fullCon
       //here reset tableForm as well when search
       dispatch({
         type: "tableForm",
-        state: { limit:10,offset:0,sortOrder:sortOrder }
+        state: { limit:10,offset:0 }
       })
     } else {
       setShowToast({ type:"warning", label: t("ES_COMMON_MIN_SEARCH_CRITERIA_MSG") })
@@ -133,17 +132,10 @@ const SearchComponent = ({ uiConfig, header = "", screenType = "search", fullCon
     //here reset tableForm as well
     dispatch({
       type: "tableForm",
-      state: { limit:10,offset:0,sortOrder:sortOrder }
+      state: { limit:10,offset:0 }
       //need to pass form with empty strings 
     })
   }
-
-  const handleSort = () => {
-    const updatedSortOrder = sortOrder === "asc" ? "desc" : "asc";
-    // Dispatch sortOrder to context
-    dispatch({ type: "updateSortOrder", state: updatedSortOrder });
-    setSortOrder((prev) => (prev === "asc" ? "desc" : "asc"));
-  };
 
   //call this fn whenever session gets updated
   const setDefaultValues = () => {
@@ -235,9 +227,9 @@ const SearchComponent = ({ uiConfig, header = "", screenType = "search", fullCon
     >
       {uiConfig?.type === "filter" ? (
         <FilterCard
-          title={t(uiConfig?.label) || t("Filter")}
-          primaryActionLabel={t(uiConfig?.primaryLabel) || ""}
-          secondaryActionLabel={t(uiConfig?.secondaryLabel) || ""}
+          title={uiConfig?.label || "Filter"}
+          primaryActionLabel={uiConfig?.primaryLabel || ""}
+          secondaryActionLabel={uiConfig?.secondaryLabel || ""}
           onPrimaryPressed={handleSubmit(onSubmit)}
           onSecondaryPressed={clearSearch}
           layoutType={"vertical"}
@@ -288,32 +280,20 @@ const SearchComponent = ({ uiConfig, header = "", screenType = "search", fullCon
                   )}
                   {uiConfig?.isPopUp && uiConfig?.primaryLabel && (
                     <Button
-                      variation={uiConfig?.primaryLabelVariation || "primary"}
+                      variation="primary"
                       label={t(uiConfig?.primaryLabel)}
                       type="submit"
                       size={"medium"}
-                      icon={uiConfig?.primaryLabelIcon || ""}
                       onClick={(e) => handleSubmit(e)}
                     />
                   )}
                   {!uiConfig?.isPopUp && uiConfig?.primaryLabel && (
                     <Button
-                      variation={uiConfig?.primaryLabelVariation || "primary"}
+                      variation="primary"
                       label={t(uiConfig?.primaryLabel)}
                       type="submit"
-                      icon={uiConfig?.primaryLabelIcon || ""}
                       size={"medium"}
                       onClick={(e) => handleSubmit(e)}
-                    />
-                  )}
-                  {uiConfig?.sortConfig && (
-                    <Button
-                      variation={uiConfig?.sortConfig?.variation}
-                      label={t(uiConfig?.sortConfig?.label)}
-                      type="button"
-                      size={"medium"}
-                      onClick={handleSort}
-                      icon={uiConfig?.sortConfig?.icon}
                     />
                   )}
                 </div>