+++ react/ui-components/src/atoms/SelectionTag.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState} from "react";
 import PropTypes from "prop-types";
 import ErrorMessage from "./ErrorMessage";
 import { useTranslation } from "react-i18next";
@@ -9,51 +9,13 @@ const SelectionTag = ({
   width,
   errorMessage,
   options,
-  optionsKey = "name",
   onSelectionChanged,
   allowMultipleSelection = true,
   selected,
-  withContainer,
-  populators = {},
+  withContainer
 }) => {
-  const { t: i18nT } = useTranslation();
-  const t = populators?.t || i18nT;
+  const { t } = useTranslation();
   const [selectedOptions, setSelectedOptions] = useState(selected || []);
-
-  const { isLoading, data } = window?.Digit?.Hooks.useCustomMDMS(
-    Digit?.ULBService?.getStateId(),
-    populators?.mdmsConfig?.moduleName,
-    [
-      {
-        name: populators?.mdmsConfig?.masterName,
-      },
-    ],
-    {
-      select: populators?.mdmsConfig?.select
-        ? createFunction(populators?.mdmsConfig?.select)
-        : (data) => {
-            const optionsData = _.get(
-              data,
-              `${populators?.mdmsConfig?.moduleName}.${populators?.mdmsConfig?.masterName}`,
-              []
-            );
-            return optionsData
-              .filter((opt) =>
-                opt?.hasOwnProperty("active") ? opt.active : true
-              )
-              .map((opt) => ({
-                ...opt,
-                name: `${Digit.Utils.locale.getTransformedLocale(opt.code)}`,
-              }));
-          },
-      enabled: populators?.mdmsConfig || populators?.mdmsV2 ? true : false,
-    },
-    populators?.mdmsv2
-  );
-
-  if (isLoading) {
-    return <Loader />;
-  }
 
   const handleOptionClick = (option) => {
     const updatedSelections = [...selectedOptions];
@@ -88,7 +50,13 @@ const SelectionTag = ({
 
   const IconRender = (iconReq, isActive) => {
     const iconFill = isActive ? primaryIconColor : secondaryIconColor;
-    return iconRender(iconReq, iconFill, "1.5rem", "1.5rem", "");
+    return iconRender(
+      iconReq,
+      iconFill,
+      "1.5rem",
+      "1.5rem",
+      ""
+    );
   };
 
   const renderOption = (option) => {
@@ -107,9 +75,7 @@ const SelectionTag = ({
             {IconRender(option?.prefixIcon, isSelected)}
           </span>
         )}
-        <span className="selectiontag-option-label">
-          {t(option?.[optionsKey])}
-        </span>
+        <span className="selectiontag-option-label">{option.name}</span>
         {option.suffixIcon && (
           <span className="selectiontagicon">
             {IconRender(option?.suffixIcon, isSelected)}
@@ -121,12 +87,8 @@ const SelectionTag = ({
 
   return (
     <div className="selection-card-container">
-      <div
-        className={`selection-card ${errorMessage ? "error" : ""} ${
-          !withContainer ? "hideContainer" : ""
-        }`}
-      >
-        {data ? data?.map(renderOption) : options?.map(renderOption)}
+      <div className={`selection-card ${errorMessage ? "error" : ""} ${!withContainer ? "hideContainer" : ""}`}>
+        {options?.map(renderOption)}
       </div>
       {errorMessage && (
         <ErrorMessage