+++ react/ui-components/src/atoms/SubmitBar.js
@@ -1,19 +1,7 @@
 import React, { forwardRef } from "react";
 import PropTypes from "prop-types";
-import { iconRender } from "../utils/iconRender";
-import { Colors } from "../constants/colors/colorconstants";
 
 const SubmitBar = forwardRef((props, ref) => {
-  const primaryIconColor = Colors.lightTheme.paper.primary;
-
-  const icon = iconRender(
-    props?.icon,
-    props?.iconFill || primaryIconColor,
-    "1.5rem",
-    "1.5rem",
-    `digit-button-customIcon`
-  );
-
   return (
     <button
       ref={ref}
@@ -21,14 +9,10 @@ const SubmitBar = forwardRef((props, ref) => {
       className={`${props.disabled ? `digit-submit-bar ${"disabled"}` : "digit-submit-bar"} ${props.className ? props.className : ""}`}
       type={props.submit ? "submit" : "button"}
       style={{ ...props.style }}
-      onClick={props.onSubmit}
+      onClick={props.submit ? props.onSubmit : props.onClick}
       {... props.form ? {form: props.form} : {}}
     >
-      <div className={`icon-label-container`}>
-        {!props?.isSuffix && props?.icon && icon}
-        <h2 className="digit-button-label">{props.label}</h2>
-        {props?.isSuffix && props?.icon && icon}
-      </div>
+      <h2>{props.label}</h2>
     </button>
   );
 });