+++ react/ui-components/src/hoc/BoundaryFilter.js
@@ -502,12 +502,11 @@ const BoundaryFilter = (props) => {
               return highestIndex <= index && index <= lowestIndex;
             })?.map((item) => {
 
-
               return (item?.boundaryType === rootBoundaryType) ? (
                 <LabelFieldPair key={item.boundaryType} className="boundary-item" vertical={!updatedLayoutConfig?.isLabelFieldLayoutHorizontal}>
-                  <CardLabel className={"boundary-selection-label"}>
+                  {updatedLayoutConfig?.isLabelNeeded && <CardLabel className={"boundary-selection-label"}>
                     {item?.boundaryType}
-                  </CardLabel>
+                  </CardLabel>}
                   <div className="digit-field-full">
                     {!(props?.levelConfig?.isSingleSelect && props?.levelConfig?.isSingleSelect?.includes(item?.boundaryType)) ?
                       <MultiSelectDropdown
@@ -575,10 +574,10 @@ const BoundaryFilter = (props) => {
 
                     return (
                       <LabelFieldPair key={item.boundaryType} className="boundary-item" vertical={!updatedLayoutConfig?.isLabelFieldLayoutHorizontal}>
-                        <CardLabel className={"boundary-selection-label"}>
+                        {updatedLayoutConfig?.isLabelNeeded  && <CardLabel className={"boundary-selection-label"}>
                           {t((hierarchyType + "_" + item?.boundaryType).toUpperCase())}
-                        </CardLabel>
-                        <div className="digit-field-full">
+                        </CardLabel>}
+                        <div className="digit-field-full" style={{width:"100%"}}>
                           {!(props?.levelConfig?.isSingleSelect && props?.levelConfig?.isSingleSelect?.includes(item?.boundaryType)) ?
                             <>
                               <MultiSelectDropdown