+++ react/ui-components/src/hoc/FormComposerV2.js
@@ -11,7 +11,8 @@ import ActionLinks from "../atoms/ActionLinks";
 import Footer from "../atoms/Footer";
 import LabelFieldPair from "../atoms/LabelFieldPair";
 import HorizontalNav from "../atoms/HorizontalNav";
-import { SubmitBar, Toast , Button } from "../atoms";
+import { SubmitBar, Toast } from "../atoms";
+import MultiChildFormWrapper from "./MultiChildFormWrapper";
 
 // import Fields from "./Fields";    //This is a field selector pickup from formcomposer
 import FieldController from "./FieldController";
@@ -68,13 +69,6 @@ export const FormComposer = (props) => {
   //clear all errors if user has changed the form category.
   //This is done in case user first click on submit and have errors in cat 1, switches to cat 2 and hit submit with errors
   //So, he should not get error prompts from previous cat 1 on cat 2 submit.
-
-  useEffect(() => {
-    if (props?.defaultValues && Object.keys(props?.defaultValues).length > 0) {
-      reset(props?.defaultValues);
-    }
-  }, [props?.defaultValues]);
-
   useEffect(() => {
     clearErrors();
   }, [selectedFormCategory]);
@@ -129,6 +123,8 @@ export const FormComposer = (props) => {
       control: control,
       props: props,
       errors: errors,
+      control:control,
+      defaultValues: props?.defaultValues,
       controllerProps: {
         register,
         handleSubmit,
@@ -186,20 +182,20 @@ export const FormComposer = (props) => {
     }
   };
 
+  const titleStyle = { color: "#505A5F", fontWeight: "700", fontSize: "16px" };
+
   const getCombinedComponent = (section) => {
     if (section.head && section.subHead) {
       return (
         <>
           <HeaderComponent
-            className={`digit-card-section-header titleStyle ${section?.sectionHeadClassName || ""}`}
+            className={`digit-card-section-header`}
+            style={props?.sectionHeadStyle ? props?.sectionHeadStyle : { margin: "5px 0px" }}
             id={section.headId}
           >
             {t(section.head)}
           </HeaderComponent>
-          <HeaderComponent 
-          id={`${section.headId}_DES`}
-          className={`sectionSubHeaderStyle ${section?.sectionSubHeadClassName}`}
-          >
+          <HeaderComponent style={titleStyle} id={`${section.headId}_DES`}>
             {t(section.subHead)}
           </HeaderComponent>
         </>
@@ -207,9 +203,7 @@ export const FormComposer = (props) => {
     } else if (section.head) {
       return (
         <>
-          <HeaderComponent className={`digit-card-section-header titleStyle ${section?.sectionHeadClassName || ""}`}
-          id={section.headId}
-          >
+          <HeaderComponent className={`digit-card-section-header`} style={props?.sectionHeadStyle ? props?.sectionHeadStyle : {}} id={section.headId}>
             {t(section.head)}
           </HeaderComponent>
         </>
@@ -230,7 +224,20 @@ export const FormComposer = (props) => {
     (section, index, array, sectionFormCategory) => (
       <React.Fragment key={index}>
         {section && getCombinedComponent(section)}
-        {section.body.map((field, index) => {
+        {section?.type === "multiChildForm" && (
+          <MultiChildFormWrapper
+            key={`multi-child-${index}`}
+            config={section}
+            control={control}
+            formData={formData}
+            setValue={setValue}
+            getValues={getValues}
+            errors={errors}
+            props={props}
+            defaultValues={props?.defaultValues}
+          />
+        )}
+        {section.type !== "multiChildForm" && section.body.map((field, index) => {
           if (field?.populators?.hideInForm) return null;
           if (props.inline)
             return (
@@ -373,6 +380,10 @@ export const FormComposer = (props) => {
     </React.Fragment>
   );
 
+  function onDraftLabelClick() {
+    props.onDraftLabelClick(getValues());
+  }
+
   return (
     <form onSubmit={handleSubmit(onSubmit)} onKeyDown={(e) => checkKeyDown(e)} id={props.formId} className={props.className}>
       {props?.showMultipleCardsWithoutNavs ? (
@@ -423,11 +434,9 @@ export const FormComposer = (props) => {
         </HorizontalNav>
       )}
       {!props.submitInForm && props.label && (
-        <Footer className={props.actionClassName}>
-          <SubmitBar label={t(props.label)} className="digit-formcomposer-submitbar" submit="submit" disabled={isDisabled} icon={props?.primaryActionIcon} isSuffix={props?.primaryActionIconAsSuffix} />
-          {props?.secondaryLabel && props?.showSecondaryLabel && (
-            <Button className="previous-button"  variation="secondary" label={t(props?.secondaryLabel)} onClick={props?.onSecondayActionClick} icon={props?.secondaryActionIcon} isSuffix={props?.secondaryActionIconAsSuffix} />
-          )}
+        <Footer>
+          {props?.draftLabel && (  <SubmitBar className="digit-formcomposer-submitbar" submit={false} label={t(props?.draftLabel)} onClick={onDraftLabelClick} />)}
+          <SubmitBar label={t(props.label)} className="digit-formcomposer-submitbar" submit="submit" disabled={isDisabled} />
           {props.onSkip && props.showSkip && <ActionLinks style={props?.skipStyle} label={t(`CS_SKIP_CONTINUE`)} onClick={props.onSkip} />}
         </Footer>
       )}