+++ react/ui-components/src/hoc/ResultsDataTableWrapper.js
@@ -7,22 +7,23 @@ import React, {
   useContext,
 } from "react";
 import { useTranslation } from "react-i18next";
+import TextInput from "../atoms/TextInput";
 import { useForm, Controller } from "react-hook-form";
 import _ from "lodash";
 import { InboxContext } from "./InboxSearchComposerContext";
-import { Loader } from "../atoms";
+import { Card,Loader } from "../atoms";
 import { CustomSVG } from "../atoms";
 import CheckBox from "../atoms/CheckBox";
 import NoResultsFound from "../atoms/NoResultsFound";
 import Button from "../atoms/Button";
+import CardLabel from "../atoms/CardLabel";
 import ResultsDataTable from "../molecules/ResultsDataTable";
-import { useHistory } from "react-router-dom";
+import { useHistory} from "react-router-dom";
 import { SVG } from "../atoms";
-import { Toast } from "../atoms";
+import {Toast} from "../atoms";
 import FieldV1 from "./FieldV1";
 import EditablePopup from "./EditablePopup";
-import TableRow from "../atoms/TableRow";
-import TableCell from "../atoms/TableCell";
+
 
 const useDebounce = (value, delay) => {
   const [debouncedValue, setDebouncedValue] = useState(value);
@@ -39,7 +40,6 @@ const useDebounce = (value, delay) => {
 };
 
 const ResultsDataTableWrapper = ({
-  tabData,
   config,
   data,
   isLoading,
@@ -51,12 +51,7 @@ const ResultsDataTableWrapper = ({
   browserSession,
   additionalConfig,
   TotalCount,
-  refetch,
-  manualPagination,
-  onNextPage,
-  onPrevPage,
-  onPageSizeChange,
-  rowsPerPageOptions = [5, 10, 15, 20],
+  refetch
 }) => {
   const { apiDetails } = fullConfig;
   const { t } = useTranslation();
@@ -379,7 +374,7 @@ const ResultsDataTableWrapper = ({
   });
 
   const formData = watch();
-
+  
   //call this fn whenever session gets updated
   const setDefaultValues = () => {
     reset(defaultValuesFromSession);
@@ -518,22 +513,6 @@ const ResultsDataTableWrapper = ({
     setValue("offset", newOffset);
     handleSubmit(onSubmit)();
   };
-
-  const handleRowsPerPageChangeThroughEvent = (event) => {
-    setRowsPerPage(Number(event?.target?.value));
-    setCurrentPage(1);
-    const newLimit = Number(event?.target?.value);
-    const newOffset = (currentPage - 1) * Number(event?.target?.value);
-    setLimitAndOffset({
-      limit: newLimit,
-      offset: newOffset,
-    });
-    setValue("limit", newLimit);
-    setValue("offset", newOffset);
-    handleSubmit(onSubmit)();
-  };
-
-
   useEffect(() => {
     if (limitAndOffset) {
       setValue("limit", limitAndOffset.limit);
@@ -541,99 +520,10 @@ const ResultsDataTableWrapper = ({
     }
   }, [limitAndOffset]);
 
-  const indexOfLastRow = currentPage * rowsPerPage;
-  const indexOfFirstRow = indexOfLastRow - rowsPerPage;
-  const totalPages = Math.ceil(
-    data?.[TotalCount] || filteredData?.length / rowsPerPage
-  );
-
   if (isLoading || isFetching) return <div className="digit-table-loader"><Loader/></div>;
   if (!data) return <></>;
   if (!showResultsTable) return <></>;
   if (searchResult?.length === 0) return <NoResultsFound />;
-  // Extract the custom component
-  const CustomRowComponent = Digit.ComponentRegistryService.getComponent(config?.customRow?.customRowComponent);
-  if (
-    config?.customRow?.overRideRowWithCustomRowComponent &&
-    CustomRowComponent
-  ) {
-    return (
-      <>
-        <div className="digit-custom-row-wrapper">
-          {filteredData.map((rowData, index) => (
-            <CustomRowComponent key={index} rowData={rowData} tabData={tabData} {...config?.customRow?.customRowProps}/>
-          ))}
-        </div>
-        <TableRow className={`footer-pagination-content ${"digit-results-table"}`}>
-          <TableCell
-            isHeader={false}
-            isFooter={true}
-          >
-            <div className="footer-content">
-              <div className={"footer-pagination-container"}>
-                <div className="rows-per-page">
-                  <label htmlFor="rowsPerPage">
-                    {t("CS_COMMON_ROWS_PER_PAGE")}
-                    {":"}{" "}
-                  </label>
-                  <select
-                    className="pagination-dropdown"
-                    id="rowsPerPage"
-                    value={rowsPerPage}
-                    onChange={
-                      manualPagination ? onPageSizeChange : handleRowsPerPageChangeThroughEvent
-                    }
-                  >
-                    {rowsPerPageOptions.map((option) => (
-                      <option key={option} value={option}>
-                        {option}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                <div>
-                  <span>
-                    {indexOfFirstRow + 1}-
-                    {Math.min(
-                      indexOfLastRow,
-                      data?.[TotalCount] || filteredData?.length
-                    )}{" "}
-                    of {data?.[TotalCount] || filteredData?.length}
-                  </span>
-                </div>
-                <div className="pagination custom-results">
-                  <button
-                    onClick={
-                      manualPagination
-                        ? () => onPrevPage()
-                        : () => handlePageChange(currentPage - 1)
-                    }
-                    disabled={currentPage === 1}
-                  >
-                    <SVG.ChevronLeft
-                      fill={currentPage === 1 ? "#C5C5C5" : "#363636"}
-                    ></SVG.ChevronLeft>
-                  </button>
-                  <button
-                    onClick={
-                      manualPagination
-                        ? () => onNextPage()
-                        : () => handlePageChange(currentPage + 1)
-                    }
-                    disabled={currentPage === totalPages}
-                  >
-                    <SVG.ChevronRight
-                      fill={currentPage === totalPages ? "#C5C5C5" : "#363636"}
-                    ></SVG.ChevronRight>
-                  </button>
-                </div>
-              </div>
-            </div>
-          </TableCell>
-        </TableRow>
-      </>
-    );
-  }
   return (
     <>
     <ResultsDataTable
@@ -660,8 +550,9 @@ const ResultsDataTableWrapper = ({
       defaultSortAsc={config?.defaultSortAsc}
       pagination={config.isPaginationRequired}
       paginationTotalRows={
-        data?.[TotalCount] ||
+        TotalCount ||
         data?.count ||
+        data?.TotalCount ||
         data?.totalCount ||
         searchResult?.length
       }