+++ react/ui-components/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.2.0-beta.34] - 2025-06-12
+## Enhancements
+- made changes that are required for DIGIT-STUDIO
+- added MultiChildFormWrapper to allow user to add multiple entries of a form section
+- added UploadAndDownloadDocumentHandler to allow upload and download document fields
+- removed type boundary from the render form fields
+
 ### [0.2.0-beta.33] - 2025-06-10
 ## Enhancements
 - added props to custom row component in results